requirements.txt
requests
beautifulsoup4
fpdf2
python-dotenv
